Fixed typing warning (implicit float to int conversion)

+++ src/Engine.cpp
@@ -295,7 +295,7 @@ namespace Smasher {
 	}
 
 	void Engine::OnWindowResize(Events::WindowResizeEvent& event) {
-		m_WindowView.setSize(sf::Vector2f(event.WindowSize.x, event.WindowSize.y));
+		m_WindowView.setSize(sf::Vector2f((float)event.WindowSize.x, (float)event.WindowSize.y));
 		m_Window->setView(m_WindowView);
 	}
